Remove desktop.
So if you have desktop turned on and disable a minor
mode (temporarily) and exit Emacs, that minor mode will be disabled in
that buffer the next time Emacs starts up. That provided for a lot of
head scratching: my init file turns the minor mode on but for some
reason it is turned off in this buffer. What I really want is just for
the file names to be saved and restored...maybe there is a way to do
this in desktop mode.